pyproject.toml
[project]
name = "podcast-transcribe-chinese"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.58.0",
    "elevenlabs>=2.39.0",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
    "socksio>=1.0.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
